Fix desktop backend warnings
Fixed warnings in SDL3 and desktop backend renderers:

- Added __attribute__((unused)) to unused functions and variables
- Fixed uninitialized variables in sdl3_backend.c
- Added (void) casts for unused parameters
- Fixed sign comparison warnings with appropriate casts
- Replaced strncpy with memcpy for safer string operations
- Added fread return value checking

Files modified:
- renderers/sdl3/sdl3_backend.c
- renderers/sdl3/sdl3_input.c
- renderers/sdl3/sdl3_fonts.c
- backends/desktop/ir_desktop_renderer.c
- backends/desktop/desktop_test_events.c
- backends/desktop/ir_to_commands.c
- backends/desktop/ir_desktop_state.c
- backends/desktop/renderers/sdl3/sdl3_layout.c
- backends/desktop/renderers/sdl3/sdl3_fonts.c
- backends/desktop/renderers/raylib/raylib_layout.c

+++ backends/desktop/renderers/raylib/raylib_layout.c
@@ -110,7 +110,7 @@ static int raylib_wrap_text_heuristic(const char* text, float max_width,
 
     size_t text_len = strlen(text);
     float char_width = font_size * TEXT_CHAR_WIDTH_RATIO;
-    int chars_per_line = max_width > 0 ? (int)(max_width / char_width) : strlen(text);
+    int chars_per_line = max_width > 0 ? (int)(max_width / char_width) : (int)strlen(text);
 
     if (chars_per_line <= 0) chars_per_line = 1;
